% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pars.R
\name{read_pars}
\alias{read_pars}
\title{Read simulation parameters}
\usage{
read_pars(
  dir,
  filename = "parameters.txt",
  sep = " ",
  is_num = "all",
  is_bool = "none"
)
}
\arguments{
\item{dir}{Path to the simulation folder}

\item{filename}{Name of the parameter file}

\item{sep}{Separator between parameter names and values}

\item{is_num, is_bool}{Either optional names of parameters to convert to
numeric or logical, respectively, or \code{"all"} for all (default of
\code{is_num}), or \code{"none"} (default of \code{is_bool}) for no
conversion. Note that conversion to logical will override conversion to
numeric.}
}
\value{
A named list of parameter values
}
\description{
Read the parameter file in a simulation folder
}
\details{
The function looks for a parameter text file in the folder, and
reads each line. The first entry of each line must be the parameter name,
and the rest of the line a list of values. The values are separated by
a given separator (default is a space). The function returns a list of
values (or vectors of values if applicable) named after their respective
parameter.
}
\examples{

# Path to a simulation folder
path <- system.file("extdata", "example", package = "readsim")

# Read parameters
read_pars(path)

# Read parameters with custom conversion
read_pars(path, is_num = c("popsize", "maxgrowths"), is_bool = "savepars")

}
