% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Write mock data}
\usage{
write_data(values, filename)
}
\arguments{
\item{values}{Vector of numeric values to write}

\item{filename}{Name of the file to write to}
}
\value{
Nothing (saves to file)
}
\description{
Write a vector of values into a binary file.
}
\note{
This function is for testing purposes mostly, and can be used to
play with the \code{read_data} function.
}
\examples{

# Write mock data
values <- rnorm(100)
write_data(1:10, "mock.dat")

# Read the data back
read_data(dir = ".", "mock")

# Delete the file
file.remove("mock.dat")

}
